Patch 8

+++ code/modules/clothing/glasses/glasses.dm
@@ -1,15 +1,15 @@
 /obj/item/clothing/glasses
 	name = "glasses"
 	icon = 'icons/obj/clothing/glasses.dmi'
-	item_icons = list(
+	worn_icon_list = list(
 		slot_l_hand_str = 'icons/mob/inhands/clothing/glasses_left.dmi',
 		slot_r_hand_str = 'icons/mob/inhands/clothing/glasses_right.dmi',
 	)
 	w_class = WEIGHT_CLASS_SMALL
 	active = TRUE
-	flags_inventory = COVEREYES
-	flags_equip_slot = ITEM_SLOT_EYES
-	flags_armor_protection = EYES
+	inventory_flags = COVEREYES
+	equip_slot_flags = ITEM_SLOT_EYES
+	armor_protection_flags = EYES
 	/// If TRUE it will help with near-sightness
 	var/prescription = FALSE
 	// If TRUE we are able to toggle the glasses and spawn with toggle action
@@ -31,11 +31,19 @@
 	///Color to use for the HUD tint; leave null if no tint
 	var/tint
 
+/obj/item/clothing/glasses/examine_descriptor(mob/user)
+	return "eyewear"
+
+/obj/item/clothing/glasses/examine_tags(mob/user)
+	. = ..()
+	if(prescription)
+		.["prescription"] = "It will help reduce symptoms of nearsightedness when worn."
+
 /obj/item/clothing/glasses/Initialize(mapload)
 	if(toggleable)
 		actions_types = list(/datum/action/item_action/toggle)
 	. = ..()
-	if(active)	//For glasses that spawn active
+	if(toggleable && active)	//For glasses that spawn active
 		active = FALSE
 		activate()
 
@@ -82,61 +90,60 @@
 	name = "science goggles"
 	desc = "The goggles do nothing! Can be used as safety googles."
 	icon_state = "purple"
-	item_state = "glasses"
+	worn_icon_state = "glasses"
 
 /obj/item/clothing/glasses/eyepatch
 	name = "eyepatch"
 	desc = "Yarr."
 	icon_state = "eyepatch"
-	item_state = "eyepatch"
-	flags_armor_protection = NONE
+	worn_icon_state = "eyepatch"
+	armor_protection_flags = NONE
 
 /obj/item/clothing/glasses/eyepatch/attackby(obj/item/I, mob/user, params)
 	. = ..()
-
+	if(.)
+		return
+	var/obj/item/clothing/glasses/eyepatch
 	if(istype(I, /obj/item/clothing/glasses/hud/health))
-		var/obj/item/clothing/glasses/hud/medpatch/P = new
+		eyepatch = new /obj/item/clothing/glasses/hud/medpatch
 		to_chat(user, span_notice("You fasten the medical hud projector to the inside of the eyepatch."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
 	else if(istype(I, /obj/item/clothing/glasses/meson))
-		var/obj/item/clothing/glasses/meson/eyepatch/P = new
+		eyepatch = new /obj/item/clothing/glasses/meson/eyepatch
 		to_chat(user, span_notice("You fasten the meson projector to the inside of the eyepatch."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
 	if(istype(I, /obj/item/clothing/glasses/night/imager_goggles))
-		var/obj/item/clothing/glasses/night/imager_goggles/eyepatch/P = new
+		eyepatch = new /obj/item/clothing/glasses/night/imager_goggles/eyepatch
 		to_chat(user, span_notice("You fasten the optical scanner to the inside of the eyepatch."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
-
-		update_icon()
+	if(!eyepatch)
+		return
+	qdel(I)
+	qdel(src)
+	user.put_in_hands(eyepatch)
+	update_icon()
 
 /obj/item/clothing/glasses/monocle
 	name = "monocle"
 	desc = "Such a dapper eyepiece!"
 	icon_state = "monocle"
-	flags_armor_protection = NONE
+	armor_protection_flags = NONE
 
 /obj/item/clothing/glasses/regular
 	name = "\improper regulation prescription glasses"
 	desc = "The Corps may call them Regulation Prescription Glasses but you know them as Rut Prevention Glasses."
 	icon_state = "glasses"
-	item_state = "glasses"
+	worn_icon_state = "glasses"
 	prescription = TRUE
 
 /obj/item/clothing/glasses/regular/attackby(obj/item/I, mob/user, params)
 	. = ..()
+	if(.)
+		return
 
 	if(istype(I, /obj/item/clothing/glasses/hud/health))
-		var/obj/item/clothing/glasses/hud/medglasses/P = new
+		var/obj/item/clothing/glasses/hud/medglasses/our_glasses = new
 		to_chat(user, span_notice("You fasten the medical hud projector to the inside of the glasses."))
 		qdel(I)
 		qdel(src)
-		user.put_in_hands(P)
+		user.put_in_hands(our_glasses)
 
 		update_icon()
 
@@ -144,78 +151,63 @@
 	name = "prescription glasses"
 	desc = "Made by Uncool. Co."
 	icon_state = "hipster_glasses"
-	item_state = "hipster_glasses"
+	worn_icon_state = "hipster_glasses"
 
 /obj/item/clothing/glasses/green
 	name = "green glasses"
 	desc = "Forest green glasses, like the kind you'd wear when hatching a nasty scheme."
 	icon_state = "green"
-	item_state = "green"
+	worn_icon_state = "green"
 
 /obj/item/clothing/glasses/mgoggles
 	name = "marine ballistic goggles"
 	desc = "Standard issue TGMC goggles. Mostly used to decorate one's helmet."
 	icon_state = "mgoggles"
-	item_state = "mgoggles"
+	worn_icon_state = "mgoggles"
 	soft_armor = list(MELEE = 40, BULLET = 40, LASER = 0, ENERGY = 15, BOMB = 35, BIO = 10, FIRE = 30, ACID = 30)
-	flags_equip_slot = ITEM_SLOT_EYES|ITEM_SLOT_MASK
+	equip_slot_flags = ITEM_SLOT_EYES|ITEM_SLOT_MASK
 	goggles_layer = TRUE
 	w_class = WEIGHT_CLASS_TINY
 
 /obj/item/clothing/glasses/mgoggles/attackby(obj/item/our_item, mob/user, params)
 	. = ..()
+	if(.)
+		return
+	var/obj/item/clothing/glasses/our_glasses
+
 	if(istype(our_item, /obj/item/clothing/glasses/night/imager_goggles))
 		if(prescription)
-			var/obj/item/clothing/glasses/night/optgoggles/prescription/our_glasses = new
-			to_chat(user, span_notice("You fasten the optical imaging scanner to the inside of the goggles."))
-			qdel(our_item)
-			qdel(src)
-			user.put_in_hands(our_glasses)
+			our_glasses = new /obj/item/clothing/glasses/night/optgoggles/prescription
 		else
-			var/obj/item/clothing/glasses/night/optgoggles/our_glasses = new
-			to_chat(user, span_notice("You fasten the optical imaging scanner to the inside of the goggles."))
-			qdel(our_item)
-			qdel(src)
-			user.put_in_hands(our_glasses)
+			our_glasses = new /obj/item/clothing/glasses/night/optgoggles
+		to_chat(user, span_notice("You fasten the optical imaging scanner to the inside of the goggles."))
 
-		update_icon(user)
+	else if(istype(our_item, /obj/item/clothing/glasses/hud/health))
+		if(prescription)
+			our_glasses = new /obj/item/clothing/glasses/hud/medgoggles/prescription
+		else
+			our_glasses = new /obj/item/clothing/glasses/hud/medgoggles
+		to_chat(user, span_notice("You fasten the medical hud projector to the inside of the goggles."))
+
+	else if(istype(our_item, /obj/item/clothing/glasses/meson))
+		if(prescription)
+			our_glasses = new /obj/item/clothing/glasses/meson/enggoggles/prescription
+		else
+			our_glasses = new /obj/item/clothing/glasses/meson/enggoggles
+		to_chat(user, span_notice("You fasten the optical meson scanner to the inside of the goggles."))
+
+	if(!our_glasses)
+		return
+
+	qdel(our_item)
+	qdel(src)
+	user.put_in_hands(our_glasses)
+	update_icon()
 
 /obj/item/clothing/glasses/mgoggles/prescription
 	name = "prescription marine ballistic goggles"
 	desc = "Standard issue TGMC goggles. Mostly used to decorate one's helmet. Contains prescription lenses in case you weren't sure if they were lame or not."
 	prescription = TRUE
-
-/obj/item/clothing/glasses/mgoggles/attackby(obj/item/I, mob/user, params)
-	. = ..()
-
-	if(istype(I, /obj/item/clothing/glasses/hud/health))
-		if(prescription)
-			var/obj/item/clothing/glasses/hud/medgoggles/prescription/P = new
-			to_chat(user, span_notice("You fasten the medical hud projector to the inside of the goggles."))
-			qdel(I)
-			qdel(src)
-			user.put_in_hands(P)
-		else
-			var/obj/item/clothing/glasses/hud/medgoggles/S = new
-			to_chat(user, span_notice("You fasten the medical hud projector to the inside of the goggles."))
-			qdel(I)
-			qdel(src)
-			user.put_in_hands(S)
-	else if(istype(I, /obj/item/clothing/glasses/meson))
-		if(prescription)
-			var/obj/item/clothing/glasses/meson/enggoggles/prescription/P = new
-			to_chat(user, span_notice("You fasten the optical meson scanner to the inside of the goggles."))
-			qdel(I)
-			qdel(src)
-			user.put_in_hands(P)
-		else
-			var/obj/item/clothing/glasses/meson/enggoggles/S = new
-			to_chat(user, span_notice("You fasten the optical meson scanner to the inside of the goggles."))
-			qdel(I)
-			qdel(src)
-			user.put_in_hands(S)
-
-		update_icon()
 
 /obj/item/clothing/glasses/m42_goggles
 	name = "\improper M42 scout sight"
@@ -231,8 +223,9 @@
 	name = "welding goggles"
 	desc = "Protects the eyes from welders, approved by the mad scientist association."
 	icon_state = "welding-g"
-	item_state = "welding-g"
+	worn_icon_state = "welding-g"
 	toggleable = TRUE
+	inventory_flags = COVEREYES
 	eye_protection = 2
 	activation_sound = null
 	deactivation_sound = null
@@ -242,7 +235,7 @@
 	AddComponent(/datum/component/clothing_tint, TINT_5, TRUE)
 
 /obj/item/clothing/glasses/welding/verb/verbtoggle()
-	set category = "Object.Clothing"
+	set category = "IC.Clothing"
 	set name = "Adjust welding goggles"
 	set src in usr
 
@@ -257,13 +250,12 @@
 		flip_up(user)
 
 	//This sends a signal that toggles the tint component's effects
-	toggle_item_state(user)
+	toggle_worn_icon_state(user)
 
 ///Toggle the welding goggles on
 /obj/item/clothing/glasses/welding/proc/flip_up(mob/user)
-	DISABLE_BITFIELD(flags_inventory, COVEREYES)
-	DISABLE_BITFIELD(flags_inv_hide, HIDEEYES)
-	DISABLE_BITFIELD(flags_armor_protection, EYES)
+	DISABLE_BITFIELD(inventory_flags, COVEREYES)
+	DISABLE_BITFIELD(armor_protection_flags, EYES)
 	eye_protection = 0
 	update_icon()
 	if(user)
@@ -271,9 +263,8 @@
 
 ///Toggle the welding goggles off
 /obj/item/clothing/glasses/welding/proc/flip_down(mob/user)
-	ENABLE_BITFIELD(flags_inventory, COVEREYES)
-	ENABLE_BITFIELD(flags_inv_hide, HIDEEYES)
-	ENABLE_BITFIELD(flags_armor_protection, EYES)
+	ENABLE_BITFIELD(inventory_flags, COVEREYES)
+	ENABLE_BITFIELD(armor_protection_flags, EYES)
 	eye_protection = initial(eye_protection)
 	update_icon()
 	if(user)
@@ -291,7 +282,7 @@
 	name = "superior welding goggles"
 	desc = "Welding goggles made from more expensive materials, strangely smells like potatoes."
 	icon_state = "rwelding-g"
-	item_state = "rwelding-g"
+	worn_icon_state = "rwelding-g"
 
 /obj/item/clothing/glasses/welding/superior/Initialize(mapload)
 	. = ..()
@@ -303,7 +294,7 @@
 	name = "sunglasses"
 	desc = "Strangely ancient technology used to help provide rudimentary eye cover. Enhanced shielding blocks many flashes."
 	icon_state = "sun"
-	item_state = "sunglasses"
+	worn_icon_state = "sunglasses"
 	eye_protection = 1
 
 /obj/item/clothing/glasses/sunglasses/Initialize(mapload)
@@ -315,7 +306,7 @@
 	name = "blindfold"
 	desc = "Covers the eyes, preventing sight."
 	icon_state = "blindfold"
-	item_state = "blindfold"
+	worn_icon_state = "blindfold"
 	eye_protection = 2
 
 /obj/item/clothing/glasses/sunglasses/blindfold/Initialize(mapload)
@@ -325,7 +316,7 @@
 /obj/item/clothing/glasses/sunglasses/big
 	desc = "Strangely ancient technology used to help provide rudimentary eye cover. Larger than average enhanced shielding blocks many flashes."
 	icon_state = "bigsunglasses"
-	item_state = "bigsunglasses"
+	worn_icon_state = "bigsunglasses"
 
 /obj/item/clothing/glasses/sunglasses/big/prescription
 	name = "prescription sunglasses"
@@ -337,33 +328,28 @@
 
 /obj/item/clothing/glasses/sunglasses/fake/attackby(obj/item/I, mob/user, params)
 	. = ..()
+	if(.)
+		return
 
+	var/obj/item/clothing/glasses/our_glasses
 	if(istype(I, /obj/item/clothing/glasses/hud/health))
-		var/obj/item/clothing/glasses/hud/medsunglasses/P = new
+		our_glasses = new /obj/item/clothing/glasses/hud/medsunglasses
 		to_chat(user, span_notice("You fasten the medical hud projector to the inside of the glasses."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
 	else if(istype(I, /obj/item/clothing/glasses/meson))
-		var/obj/item/clothing/glasses/meson/sunglasses/P = new
+		our_glasses = new /obj/item/clothing/glasses/meson/sunglasses
 		to_chat(user, span_notice("You fasten the optical meson scaner to the inside of the glasses."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
 	else if(istype(I, /obj/item/clothing/glasses/night/m56_goggles))
-		var/obj/item/clothing/glasses/night/sunglasses/P = new
+		our_glasses = new /obj/item/clothing/glasses/night/sunglasses
 		to_chat(user, span_notice("You fasten the KTLD sight to the inside of the glasses."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
 	else if(istype(I, /obj/item/clothing/glasses/night/imager_goggles))
-		var/obj/item/clothing/glasses/night/imager_goggles/sunglasses/P = new
+		our_glasses = new /obj/item/clothing/glasses/night/imager_goggles/sunglasses
 		to_chat(user, span_notice("You fasten the optical imager scaner to the inside of the glasses."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
-
-		update_icon()
+	if(!our_glasses)
+		return
+	qdel(I)
+	qdel(src)
+	user.put_in_hands(our_glasses)
+	update_icon()
 
 /obj/item/clothing/glasses/sunglasses/fake/prescription
 	name = "prescription sunglasses"
@@ -373,7 +359,7 @@
 	name = "designer sunglasses"
 	desc = "A pair of larger than average designer sunglasses. Doesn't seem like it'll block flashes."
 	icon_state = "bigsunglasses"
-	item_state = "bigsunglasses"
+	worn_icon_state = "bigsunglasses"
 
 /obj/item/clothing/glasses/sunglasses/fake/big/prescription
 	name = "prescription designer sunglasses"
@@ -393,95 +379,82 @@
 
 /obj/item/clothing/glasses/sunglasses/sa/nodrop
 	desc = "Glasses worn by a spatial agent. cannot be dropped"
-	flags_item = DELONDROP
+	item_flags = DELONDROP
 
 /obj/item/clothing/glasses/sunglasses/sechud
 	name = "HUDSunglasses"
 	desc = "Sunglasses with a HUD."
 	icon_state = "sunhud"
-	var/hud_type = DATA_HUD_SECURITY_ADVANCED
 
-/obj/item/clothing/glasses/sunglasses/sechud/equipped(mob/living/carbon/human/user, slot)
-	if(slot == SLOT_GLASSES)
-		var/datum/atom_hud/H = GLOB.huds[hud_type]
-		H.add_hud_to(user)
-	..()
-
-/obj/item/clothing/glasses/sunglasses/sechud/dropped(mob/living/carbon/human/user)
-	if(istype(user))
-		if(src == user.glasses) //dropped is called before the inventory reference is updated.
-			var/datum/atom_hud/H = GLOB.huds[hud_type]
-			H.remove_hud_from(user)
-	..()
+/obj/item/clothing/glasses/sunglasses/sechud/mp/Initialize()
+	. = ..()
+	AddComponent(/datum/component/clothing_tint, TINT_NONE)
 
 /obj/item/clothing/glasses/sunglasses/aviator
 	name = "aviator sunglasses"
 	desc = "A pair of aviator sunglasses."
 	icon_state = "aviator"
-	item_state = "aviator"
+	worn_icon_state = "aviator"
 
 /obj/item/clothing/glasses/sunglasses/aviator/yellow
 	name = "aviator sunglasses"
 	desc = "A pair of aviator sunglasses. Comes with yellow lens."
 	icon_state = "aviator_yellow"
-	item_state = "aviator_yellow"
+	worn_icon_state = "aviator_yellow"
 
 /obj/item/clothing/glasses/orange
 	name = "orange glasses"
 	desc = "A pair of orange glasses."
 	icon_state = "orange"
-	item_state = "orange"
+	worn_icon_state = "orange"
 	species_exception = list(/datum/species/robot)
 
 /obj/item/clothing/glasses/orange/attackby(obj/item/our_item, mob/user, params)
 	. = ..()
+	if(.)
+		return
+	var/obj/item/clothing/glasses/our_glasses
 	if(istype(our_item, /obj/item/clothing/glasses/hud/health))
-		var/obj/item/clothing/glasses/hud/orange_glasses/our_glasses = new
+		our_glasses = new /obj/item/clothing/glasses/hud/orange_glasses
 		to_chat(user, span_notice("You fasten the medical hud projector to the inside of the glasses."))
-		qdel(our_item)
-		qdel(src)
-		user.put_in_hands(our_glasses)
-		update_icon()
 	else if(istype(our_item, /obj/item/clothing/glasses/night/imager_goggles))
-		var/obj/item/clothing/glasses/night/imager_goggles/orange_glasses/our_glasses = new
+		our_glasses = new /obj/item/clothing/glasses/night/imager_goggles/orange_glasses
 		to_chat(user, span_notice("You fasten the optical imager scaner to the inside of the glasses."))
-		qdel(our_item)
-		qdel(src)
-		user.put_in_hands(our_glasses)
-		update_icon()
 	else if(istype(our_item, /obj/item/clothing/glasses/meson))
-		var/obj/item/clothing/glasses/meson/orange_glasses/our_glasses = new
+		our_glasses = new /obj/item/clothing/glasses/meson/orange_glasses
 		to_chat(user, span_notice("You fasten the optical meson scaner to the inside of the glasses."))
-		qdel(our_item)
-		qdel(src)
-		user.put_in_hands(our_glasses)
-		update_icon()
+	if(!our_glasses)
+		return
+	qdel(our_item)
+	qdel(src)
+	user.put_in_hands(our_glasses)
+	update_icon()
 
 /obj/item/clothing/glasses/meson/orange_glasses
 	name = "Orange glasses"
 	desc = "A pair of orange glasses. This pair has been fitted with an optical meson scanner."
-	item_icons = list(
+	worn_icon_list = list(
 		slot_glasses_str = 'icons/mob/clothing/eyes.dmi')
 	icon_state = "meson_orange"
-	item_state = "meson_orange"
+	worn_icon_state = "meson_orange"
 	deactive_state = "deactivated_orange"
 
 /obj/item/clothing/glasses/night/imager_goggles/orange_glasses
 	name = "Orange glasses"
 	desc = "A pair of orange glasses. This pair has been fitted with an internal optical imager scanner."
-	item_icons = list(
+	worn_icon_list = list(
 		slot_glasses_str = 'icons/mob/clothing/eyes.dmi')
 	icon_state = "optical_orange"
-	item_state = "optical_orange"
+	worn_icon_state = "optical_orange"
 	deactive_state = "deactivated_orange"
 
 /obj/item/clothing/glasses/hud/orange_glasses
 	name = "Orange glasses"
 	desc = "A pair of orange glasses. This pair has been fitted with an internal HealthMate HUD projector."
-	item_icons = list(
+	worn_icon_list = list(
 		slot_glasses_str = 'icons/mob/clothing/eyes.dmi')
 	icon_state = "med_orange"
-	item_state = "med_orange"
+	worn_icon_state = "med_orange"
 	deactive_state = "deactivated_orange"
 	toggleable = TRUE
 	hud_type = DATA_HUD_MEDICAL_ADVANCED
